put prompt in one file
